requirements.txt
aniso8601==4.0.1
asn1crypto==0.24.0
attrs==18.2.0
awscli==1.16.45
backcall==0.1.0
bleach==3.0.2
botocore==1.12.35
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.3.9
conda==4.5.11
cryptography==2.3.1
cryptography-vectors==2.3.1
cycler==0.10.0
Cython==0.29
dask==0.20.0
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
Flask==1.0.2
Flask-RESTful==0.3.6
idna==2.7
imageio==2.3.0
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==3.0.0a3
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nltk==3.3
notebook==5.7.0
numpy==1.15.3
olefile==0.46
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pyasn1==0.4.4
pycocotools==2.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
Pygments==2.2.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyrsistent==0.14.5
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.20.0
rsa==3.4.2
ruamel-yaml==0.15.71
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplejson==3.16.1
six==1.11.0
statsmodels==0.9.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==0.4.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typing==3.6.4
urllib3==1.23
visdom==0.1.8.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
